requirements.txt
colorama==0.3.9
decorator==4.2.1
Django==2.0.2
django-analytical==2.4.0
django-captcha-admin==0.2
django-filter==1.1.0
django-markdownx==2.0.22
django-recaptcha==1.3.0
djangorestframework==3.7.7
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.1
Markdown==2.6.11
parso==0.1.1
pickleshare==0.7.4
Pillow==5.0.0
prompt-toolkit==1.0.15
Pygments==2.2.0
pytz==2018.3
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
wcwidth==0.1.7
